DEVX-156 split tests
